feat: add ts code
